+++ src/main/java/com/klarna/hiverunner/HiveServerContainer.java
@@ -1,5 +1,24 @@
 /**
+ * Copyright (C) 2013-2021 Klarna AB
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/**
+<<<<<<< HEAD
  * Copyright (C) 2013-2019 Klarna AB
+=======
+ * Copyright (C) 2013-2021 Klarna AB
+>>>>>>> 2f09e0d... Ignore System.out.close calls so IntelliJ doesn't incorrectly mark tests as terminated (#146)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,6 +39,7 @@ import com.google.common.base.Joiner;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Iterables;
 import com.klarna.hiverunner.builder.Statement;
+import com.klarna.hiverunner.io.IgnoreClosePrintStream;
 import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.hadoop.hive.conf.HiveVariableSource;
 import org.apache.hadoop.hive.conf.VariableSubstitution;
@@ -35,6 +55,7 @@ import org.apache.hive.service.server.HiveServer2;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.PrintStream;
 import java.nio.file.Path;
 import javax.annotation.Nullable;
 import java.util.ArrayList;
@@ -116,11 +137,14 @@ public class HiveServerContainer {
     }
 
     public List<Object[]> executeStatement(String hiveql) {
+        // This PrintStream hack can be removed if/when IntelliJ fixes https://youtrack.jetbrains.com/issue/IDEA-120628
+        // See https://github.com/klarna/HiveRunner/issues/94 for more info.
+        PrintStream initialPrintStream = System.out;
         try {
+            System.setOut(new IgnoreClosePrintStream(System.out));
             OperationHandle handle = client.executeStatement(sessionHandle, hiveql, new HashMap<>());
             List<Object[]> resultSet = new ArrayList<>();
             if (handle.hasResultSet()) {
-
                 /*
                  * fetchResults will by default return 100 rows per fetch (hive 14). For big result sets we need to continuously fetch the result set until all
                  * rows are fetched.
@@ -146,6 +170,8 @@ public class HiveServerContainer {
         } catch (HiveSQLException e) {
             throw new IllegalArgumentException("Failed to executeQuery Hive query " + hiveql + ": " + e.getMessage(),
                     e);
+        } finally {
+            System.setOut(initialPrintStream);
         }
     }